feat: add very inefficient polling sync.

+++ src/lib/storage.ts
@@ -1,6 +1,22 @@
-import type { Agent } from "@atproto/api";
+import { Agent } from "@atproto/api";
 import type { AutodocStorageInterface } from "./autodoc.svelte";
 import * as base64 from "js-base64";
+import type { DidDocument } from "@atproto/oauth-client-browser";
+
+const handleCache: { [did: string]: DidDocument } = {};
+
+export async function resolveDid(
+  did: string,
+): Promise<DidDocument | undefined> {
+  if (handleCache[did]) return handleCache[did];
+  try {
+    const resp = await fetch(`https://plc.directory/${did}`);
+    const json = await resp.json();
+    return json;
+  } catch (_e) {
+    // Ignore error
+  }
+}
 
 /** Takes a storage adapter and creates a sub-adapter by with the given namespace. */
 export function namespacedSubstorage(
@@ -33,6 +49,11 @@ export function namespacedSubstorage(
 export class RoomyPdsStorageAdapter implements AutodocStorageInterface {
   agent: Agent;
 
+  /** Additional DIDs that records will be _downloaded from_. Obviously we cannot delete or change
+   * the records, but they can contribute to our synchronization. */
+  additionalDids: string[];
+  additionalAgents: Promise<{ [did: string]: Agent }> = Promise.resolve({});
+
   static async buildKey(key: string[]): Promise<string> {
     if (key.some((x) => x.includes("\0")))
       throw "Cannot encode paths containing null bytes";
@@ -43,19 +64,37 @@ export class RoomyPdsStorageAdapter implements AutodocStorageInterface {
     return base64.fromUint8Array(hash, true);
   }
 
-  constructor(agent: Agent) {
+  constructor(agent: Agent, ...additionalDids: string[]) {
     this.agent = agent;
+    this.additionalDids = additionalDids;
+    this.additionalAgents = (async () => {
+      const agents: { [did: string]: Agent } = {};
+      for (const did of additionalDids) {
+        const didDoc = await resolveDid(did);
+        if (!didDoc) throw `Could not resolve DID doc for ${did}`;
+        const pdsService = (didDoc.service || []).find(
+          (x) => x.id === "#atproto_pds",
+        );
+        if (!pdsService || typeof pdsService.serviceEndpoint !== "string")
+          throw `Could not resolve PDS service for ${did}`;
+        agents[did] = new Agent(pdsService.serviceEndpoint);
+      }
+
+      return agents;
+    })();
   }
 
   static async *listRecords(
     agent: Agent,
+    did: string,
     prefix: string[],
-  ): AsyncIterable<{ key: string[]; blobCid: string }> {
+  ): AsyncIterable<{ key: string[]; blobCid: string; did: string }> {
     let cursor: string | undefined;
+
     do {
       const resp = await agent.com.atproto.repo.listRecords({
         collection: "town.muni.roomy.v0.store",
-        repo: agent.assertDid,
+        repo: did,
         cursor,
         limit: 100,
       });
@@ -79,6 +118,7 @@ export class RoomyPdsStorageAdapter implements AutodocStorageInterface {
           yield {
             blobCid: rdata.data.ref.toString(),
             key: rdata.key,
+            did: did,
           };
         }
       }
@@ -127,33 +167,46 @@ export class RoomyPdsStorageAdapter implements AutodocStorageInterface {
     key: string[],
   ): Promise<{ key: string[]; data: Uint8Array | undefined }[]> {
     const records: { key: string[]; data: Uint8Array | undefined }[] = [];
-    for await (const record of RoomyPdsStorageAdapter.listRecords(
-      this.agent,
-      key,
-    )) {
-      const resp = await this.agent.com.atproto.sync.getBlob({
-        cid: record.blobCid,
-        did: this.agent.assertDid,
-      });
-      if (!resp.success) console.warn("Error downloading blob", resp);
-      records.push({
-        key: record.key,
-        data: resp.success ? new Uint8Array(resp.data.buffer) : undefined,
-      });
+
+    for (const did of [this.agent.assertDid, ...this.additionalDids]) {
+      const a =
+        did == this.agent.assertDid
+          ? this.agent
+          : (await this.additionalAgents)[did];
+
+      for await (const record of RoomyPdsStorageAdapter.listRecords(
+        a,
+        did,
+        key,
+      )) {
+        const resp = await a.com.atproto.sync.getBlob({
+          cid: record.blobCid,
+          did,
+        });
+        if (!resp.success) console.warn("Error downloading blob", resp);
+        records.push({
+          key: record.key,
+          data: resp.success ? new Uint8Array(resp.data.buffer) : undefined,
+        });
+      }
     }
+
     return records;
   }
   async removeRange(key: string[]): Promise<void> {
     for await (const record of RoomyPdsStorageAdapter.listRecords(
       this.agent,
+      this.agent.assertDid,
       key,
     )) {
-      const resp = await this.agent.com.atproto.repo.deleteRecord({
-        repo: this.agent.assertDid,
-        collection: "town.muni.roomy.v0.store",
-        rkey: await RoomyPdsStorageAdapter.buildKey(record.key),
-      });
-      if (!resp.success) console.warn("Error deleting record", resp.data);
+      if (record.did == this.agent.assertDid) {
+        const resp = await this.agent.com.atproto.repo.deleteRecord({
+          repo: this.agent.assertDid,
+          collection: "town.muni.roomy.v0.store",
+          rkey: await RoomyPdsStorageAdapter.buildKey(record.key),
+        });
+        if (!resp.success) console.warn("Error deleting record", resp.data);
+      }
     }
   }
 }
